MatchFightBackend/requirements.txt
asgiref==3.8.1
Django==4.2.3
gunicorn
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==26.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
psycopg2-binary
psycopg2
